refactoring mam tests
